recover proxy config mechanism
